+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 2.1.6 (unreleased)
 ------------------
 
+-   Catch additional non-blocking errors in the email notice validator.
+
 -   Lower omegascan minimum frequency to 10 Hz from 20 Hz to better
     frequencies used by searches.